fix: applied requested changes
