+++ src/main/kotlin/app/revanced/patches/youtube/misc/fix/backtoexitgesture/annotation/FixBackToExitGestureCompatibility.kt
@@ -1,4 +1,4 @@
-package app.revanced.patches.youtube.misc.fix.backtoexit.annotation
+package app.revanced.patches.youtube.misc.fix.backtoexitgesture.annotation
 
 import app.revanced.patcher.annotation.Compatibility
 import app.revanced.patcher.annotation.Package
@@ -10,4 +10,4 @@ import app.revanced.patcher.annotation.Package
 )
 @Target(AnnotationTarget.CLASS)
 @Retention(AnnotationRetention.RUNTIME)
-internal annotation class BackToExitCompatibility
+internal annotation class FixBackToExitGestureCompatibility
